move name parser
inline with reference parser namespacing
