+++ lib/court_data_adaptor.rb
@@ -11,6 +11,6 @@ require_relative 'court_data_adaptor/resource/defendant'
 require_relative 'court_data_adaptor/query/base'
 require_relative 'court_data_adaptor/query/defendant/by_name'
 require_relative 'court_data_adaptor/query/defendant/by_reference'
+require_relative 'court_data_adaptor/query/defendant/name_parser'
 require_relative 'court_data_adaptor/query/defendant/reference_parser'
 require_relative 'court_data_adaptor/query/prosecution_case'
-require_relative 'court_data_adaptor/name_parser'
